+++ src/commands/delete.ts
@@ -1,5 +1,6 @@
 import fs from "fs"
 import { Arguments } from "yargs"
+import { OrigeenError } from "../commands"
 
 import { h1 } from "../logs"
 import { deleteProject, getProjects } from "../projects"
@@ -12,7 +13,9 @@ export function execute(args: Arguments) {
 
     if (projectName == undefined) {
         error(`You did not provide any project to delete`)
-        return
+        throw new OrigeenError("You did not provide any project name", [
+            "Provide a <projectName> after the command name"
+        ])
     }
 
     debug(`Searching for project named '${projectName}'`)
@@ -21,7 +24,10 @@ export function execute(args: Arguments) {
         error(`A project named \`${projectName}\` wasn't found.`)
         error(`Did you misspell it?`)
         error()
-        return
+        throw new OrigeenError(`Origeen did not found any project named '${projectName}'`, [
+            "Check the spelling",
+            "List all your projects with 'orgn projects'"
+        ])
     }
     debug("Found one !")
     debug()